+++ Persistence/src/main/java/by/karelin/persistence/repositories/interfaces/IRatingRepository.java
@@ -4,5 +4,5 @@ import by.karelin.domain.models.Rating;
 
 public interface IRatingRepository {
     Rating getUserRating(Long userId, Long filmId);
-    Long setRating(int rate, Long filmId, Long userId);
+    Double setRating(int rate, Long filmId, Long userId);
 }
